Move ask_yes_or_no from experiment to utils
