+++ lagom/experiment/run_experiment.py
@@ -7,26 +7,7 @@ from datetime import timedelta
 
 from lagom.utils import color_str
 from lagom.utils import yaml_dump
-
-
-def ask_yes_or_no(msg):
-    r"""Ask user to enter yes or no to a given message. 
-    
-    Args:
-        msg (str): a message
-    """
-    
-    print(msg)
-    
-    while True:
-        answer = str(input('>>> ')).lower().strip()
-        
-        if answer[0] == 'y':
-            return True
-        elif answer[0] == 'n':
-            return False
-        else:
-            print("Please answer 'yes' or 'no':")
+from lagom.utils import ask_yes_or_no
 
 
 def run_experiment(worker_class, master_class, num_worker):
@@ -100,6 +81,4 @@ def run_experiment(worker_class, master_class, num_worker):
             
     # Run experiment in parallel
     experiment()
-
-    msg = color_str(f'\nTotal time: {timedelta(seconds=round(time() - t))}', 'green', 'bold')
-    print(msg)
+    print(color_str(f'\nTotal time: {timedelta(seconds=round(time() - t))}', 'green', 'bold'))
