+++ lagom/utils/__init__.py
@@ -9,3 +9,5 @@ from .serialize import pickle_load
 from .serialize import pickle_dump
 from .serialize import yaml_load
 from .serialize import yaml_dump
+
+from .yes_no import ask_yes_or_no
